infra: add redundancy
Set a number of replicas for `blackboards` and the two `opentracker`
images as the default number of replicas is one. This means that if a
pod fails for some reason, there will be a backup one to use for the
time being and Kubernetes will spin up another image while the backup
handles traffic.