feat: add secrets for the 2022 elections (#5)
As the Barbell elections for 2022 are coming up, we need to make some
changes to the environment variables. Mainly this involves updating the
list of members to be accurate, as provided by Angus. This also involves
setting the `PRESIDENT_ID` value to be that of Angus', in the unlikely
case of ties during the election. It also updates the `ELECTION_ADMINS`
to be Angus, Jess, Jordan and myself (just in case).