feat: add additional listener
With #20, there is also `starling-webhooks.blackboards.pl` as a DNS
record, which is where the existing traffic is being sent. For now, the
traffic for this record should be routed to the same service.

Once the webhook URL has been updated, this can be removed alongside the
DNS record, but this will allow for no downtime.