feat: add `opl-docs` instance
The other instance has an old SSH key attached to it and adding a new
one is apparently a resource consuming operation (meaning the instance
will be destroyed and recreated). That's not ideal, so let's spin up
another instance with the new SSH key and migrate things over.

This change:
* Adds the public key for the new laptop
* Adds a new droplet and DNS record at `opl-docs.blackboards.pl`