feat: create Postgres instance (#65)
The fact that the data for OpenTracker is stored on the same server as
the backend and frontend is quite annoying. This makes it very difficult
to replace the instance, since the data must remain intact and cannot be
lost.

Ideally, we'd move the data to a separate instance and begin backing it
up regularly. This would allow the stateless Docker images to be moved
around much more easily as well as torn down whenever.

This change:
* Adds a new droplet of the smallest possible size to act as a Postgres
  server