feat: add EC2 instance (#74)
In order to get some stuff running, we'll probably want an EC2 instance.
This will be a simple web server that runs `f2` and any other
applications controlled by it. While this could be done on Lambda or
something else, this is simpler for the time being.

Note: most of this has been applied locally so the plan will be largely
empty. There is however a permadiff in the `aws_instance` definition
that causes it to be recreated every time. This needs fixing.

This change:
* Adds a new bucket for `f2` configuration
* Adds a new VPC and subnet
* Adds security groups and rules to allow inbound/outbound traffic
* Adds an SSH key for the instance
* Adds the instance itself with an elastic IP
* Adds an internet gateway, route tables and associations