fix: remove permadiff
Currently, Terraform always reports that the EC2 instance has changed
and thus needs to be recreated. This is fine while it's not doing
anything but would be a pain if it was actually serving any traffic.

Turns out this is due to the use of `security_groups` despite the
instance being in a non-default VPC, so we need to correct that.

This change:
* Updates the parameter name