feat: get `f2` working on instance (#80)
In order for `f2` to be processing anything, we'll need a DNS rule to
send traffic to the server. This will allow `certbot` to generate a
valid certificate for it to use. As well as this, we'll need to allow
inbound HTTPS traffic to the instance.

Note: this has been applied locally.

This change:
* Adds the DNS rule
* Allows inbound HTTPS