feat: add module for `f2` (#86)
Terraform is designed to be composable, but the current set of
definitions is just a big ball of spaghetti. There's an `s3-bucket`
module which is useful, but it would also be good to have a module for
creating instances running `f2` for the purposes of upgrades.

Additionally, there's some user setup required at the moment to install
Docker, update all the packages and then run the `f2` container. This
can all be automated instead.

This change:
* Adds a module to create IAM roles, security groups and an EC2 instance
  running a specific version of the `f2` container
* Creates a new instance using the module, with the expectation of
  failing it over to be the primary