refactor: remove redundant rules (#89)
The AWS documentation specifies that security groups are stateful, such
that if you allow inbound or outbound traffic for something then the
return traffic is automatically allowed. This means that the ephemeral
rules we have in place are not needed.

This also made things a little confusing, there's no specific rule to
allow the instance to talk to the Digital Ocean Postgres server yet the
functionality works. This is because the ephemeral rules are allowing
all outbound traffic on high ports (of which Postgres is 5432).

This change:
* Removes the ephemeral rules
* Adds a new Postgres-specific rule to allow traffic only to the Digital
  Ocean server