refactor: allow specifying config location (#104)
Some of the S3 buckets will need recreating as part of a refactor, as
they use hardcoded identifiers instead of the `random_id` resource
native to Terraform. As part of this, a new configuration bucket will be
created, everything copied over and a new `f2-instance` spun up.

Specifying the configuration location is also useful for running a
demo/prod setup in the future or for doing failovers of `f2` itself.

Note: this is a no-op, the configuration comes out the same.

This change:
* Adds variables to the `f2-instance` module and passes them into the
  script
* Updates the instance definition