feat: create secondary instance (#106)
* feat: create secondary instance

The new configuration bucket has manually been populated with the same
files as the old one, so let's bring up an `f2` instance running against
the new one.

This change:
* Creates a new instance of the module

* fix: amend instance name

This will cause a clash in the `aws_iam_role` definitions, so let's give
this a different name for now.