feat: add SES setup (#132)
Emails for OpenTracker are currently sent via Gmail which tends to put
them in the Spam folder as the email is rarely used. It also doesn't
come from the domain itself, which makes it a little more suspicious.

SES should be relatively cheap (1000 emails a month for 10 cents) and
should allow us to use `noreply@opentracker.app`.

This change:
* Defines some infrastructure for the SES setup