feat: remove EIP for Postgres
Elastic IPs will cost money in the future and this was only so that we
could get onto the instance for setup and debugging purposes. The
instance should function fine from now, so we can remove the EIP
allocation.

We can also always reassociate it if we need to SSH onto it.

This change:
* Removes the `aws_eip` resource from the Postgres instance