fix: add IPv6 subnet CIDR
The update of the subnet failed since it didn't have an IPv6 CIDR block
to work with, so this derives it from the VPC itself using some
Terraform magic.

This will also need to finish creating the `f2-instance` from the
previous change as some of that relied on the subnet.

This change:
* Adds the `ipv6_cidr_block` parameter for the subnet