feat: remove Digital Ocean references (#168)
Since we have now moved everything into AWS, we can remove the remaining
references to Digital Ocean.

The removal of the project did cause a pipeline failure as it was the
default project, but `terraform state rm
digitalocean_project.blackboards` has caused Terraform to forget about
it for now.

This change:
* Removes it from the README
* Removes the providers and versions
* Removes the empty `digital_ocean.tf` file