feat: add SNS topic notifications
It would be good to have some monitoring of the server (even something
basic, such as "hey, the server is down"). The RustNation conference
provided a good starting point for this in the AWS Lambda talk, so let's
create an SNS topic we can publish to and route emails to myself for any
failures.

This change:
* Adds an SNS topic
* Subscribes my email address to the topic