feat: support pulling images from ECR (#201)
Since `ticket-tracker` is defined as an ECR repository (and likely
everything will be in the future), we'll need to allow the `f2-instance`
module to pull images directly from there. This is actually more
performant as well (since we're in the same region) so means starting
new containers is a little quicker.

This change:
* Adds configuration to allow `f2-instance` resources to pull images
  from the `ticket-tracker` repository
* Updates the user data script to login to ECR on startup
* Sets the module to not restart on user data changes
* Creates a new instance so it has the new login flow