feat: create logging bucket
Currently, all the logs produced by the system are entirely ephemeral.
To actually view anything, you need to go onto the instance and run
`docker logs` presuming the container is still there.

Instead, we should be archiving all the logs to persistent storage such
that they can be viewed historically if needed. This will also open up
the ability to view them without being on the server hopefully.

The plan is to use `vector` for these and write them all to S3. A test
instance (`secondary`) was spun up locally using `terraform apply` to
test this and successfully created some log files from the Docker source
into the bucket here.

This change:
* Creates the S3 bucket
* Gives `f2-instance` permissions on it
* Tears down the `secondary` instance