feat: swap instances over (#236)
The current instance seemed to have degraded performance (significantly,
considering it was unable to respond to requests entirely at times)
despite the fact the CPU performance seemed alright and the database
queries were running in sub-millisecond response times.

This has been failed over locally (due to lack of desire to make
multiple PRs), so this will be a no-op.

This change:
* Swaps over to a new instance