feat: deploy Vector configuration
The `vector.yaml` being used at the moment was manually placed in the
bucket instead of uploaded through the configuration deployer. There
haven't been any changes to it, but for when there are it would be good
to have it in source control.

Additionally, this prevents having to download the file from the bucket
to make changes.

This change:
* Creates a new `configuration` directory and moves the `f2` folder
  under it
* Adds a `vector` section under it with the existing configuration
* Updates the deployment workflow to sync the contents of this directory
  with the bucket (but not delete anything already there, as there's
  existing certificates)