feat: tidy up `uptime` resources (#251)
Now that `uptime` has been moved to running on the `f2` instances, we
can tear down a lot of the resources that were being used previously.
We'll still want the SNS topic for the future and the bucket will need
cleaning up of objects first but everything else can be removed.

This change:
* Removes the ECR repository, lambda function and trigger
* Marks the bucket for deletion