feat: add support for telemetry
The `telemetry` instance runs ClickStack (or some of the components of
it) and allows traces to be exported from the main `f2` instances into
it. Additionally, it's exposed behind mTLS so we can poke around at the
data.

This has already been applied locally through a lot of hacking around...

This change:
* Aligns the Terraform with what is actually being used